+++ StrategyMapGenerationPlugin/tests/RandomGeneratorTests.cpp
@@ -4,29 +4,29 @@
 
 TEST(RandomGeneratorTest, DeterministicRandomGeneration) {
     RandomGenerator randomGenerator(1234);
-    std::list<float> a = randomGenerator.GenerateListBetween(0, 1, 5);
+    std::list<int> a = randomGenerator.GenerateListBetween(0, 1, 5);
 
     RandomGenerator randomGenerator2(1234);
-    std::list<float> b = randomGenerator2.GenerateListBetween(0, 1, 5);
+    std::list<int> b = randomGenerator2.GenerateListBetween(0, 1, 5);
 
     ASSERT_EQ(a, b);
 }
 
 TEST(RandomGeneratorTest, DivergentRandomGeneration) {
     RandomGenerator randomGenerator(1234);
-    std::list<float> a = randomGenerator.GenerateListBetween(0, 1, 5);
+    std::list<int> a = randomGenerator.GenerateListBetween(0, 1, 5);
 
     RandomGenerator randomGenerator2(5678);
-    std::list<float> b = randomGenerator2.GenerateListBetween(0, 1, 5);
+    std::list<int> b = randomGenerator2.GenerateListBetween(0, 1, 5);
 
     ASSERT_NE(a, b);
 }
 
 TEST(RandomGeneratorTest, ValueRangeVerification) {
     RandomGenerator randomGenerator(1234);
-    std::list<float> values = randomGenerator.GenerateListBetween(5.0f, 10.0f, 1000);
+    std::list<int> values = randomGenerator.GenerateListBetween(5, 10, 1000);
 
-    for (float value : values) {
+    for (int value : values) {
         ASSERT_GE(value, 5.0f);
         ASSERT_LE(value, 10.0f);
     }
@@ -34,23 +34,23 @@ TEST(RandomGeneratorTest, ValueRangeVerification) {
 
 TEST(RandomGeneratorTest, CorrectListSize) {
     RandomGenerator randomGenerator(1234);
-    std::list<float> values = randomGenerator.GenerateListBetween(0, 1, 100);
+    std::list<int> values = randomGenerator.GenerateListBetween(0, 1, 100);
 
     ASSERT_EQ(values.size(), 100);
 }
 
 TEST(RandomGeneratorTest, EmptyListGeneration) {
     RandomGenerator randomGenerator(1234);
-    std::list<float> values = randomGenerator.GenerateListBetween(0, 1, 0);
+    std::list<int> values = randomGenerator.GenerateListBetween(0, 1, 0);
 
     ASSERT_TRUE(values.empty());
 }
 
 TEST(RandomGeneratorTest, SingleValueRange) {
     RandomGenerator randomGenerator(1234);
-    std::list<float> values = randomGenerator.GenerateListBetween(5.0f, 5.0f, 100);
+    std::list<int> values = randomGenerator.GenerateListBetween(5, 5, 100);
 
-    for (float value : values) {
+    for (int value : values) {
         ASSERT_FLOAT_EQ(value, 5.0f);
     }
 }
